requirements.txt
anytree==2.8.0
argcomplete==1.10.3
beautifulsoup4==4.8.2
bitarray==2.7.3
branca==0.6.0
Brotli==1.0.9
certifi==2022.12.7
cffi==1.15.1
chardet==3.0.4
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
compressed-rtf==1.0.6
cssselect2==0.7.0
decorator==5.1.1
docx==0.2.4
docx2pdf==0.1.8
docx2txt==0.8
ebcdic==1.1.1
extract-msg==0.28.7
folium==0.14.0
fonttools==4.39.3
fpdf==1.7.2
future==0.18.3
geocoder==1.38.1
html5lib==1.1
idna==3.4
IMAPClient==2.1.0
Jinja2==3.1.2
lxml==4.9.2
MarkupSafe==2.1.2
MouseInfo==0.1.3
numpy==1.24.3
olefile==0.46
opencv-python==4.7.0.72
packaging==23.1
pandas==2.0.0
pathlib==1.0.1
pdf2image==1.16.3
pdfminer.six==20191110
pdfreader==0.1.12
Pillow==9.5.0
PyAutoGUI==0.9.53
pycparser==2.21
pycryptodome==3.17
pydyf==0.6.0
PyGetWindow==0.0.9
PyMsgBox==1.0.9
pyperclip==1.8.2
pyphen==0.14.0
PyRect==0.2.0
PyScreeze==0.1.28
pytesseract==0.3.10
python-dateutil==2.8.2
python-docx==0.8.11
python-pptx==0.6.21
python3-xlib==0.15
PythonTurtle==0.3.2
pytube==12.1.3
pytweening==1.0.4
pytz==2023.3
pytz-deprecation-shim==0.1.0.post0
ratelim==0.1.6
requests==2.28.2
six==1.12.0
sortedcontainers==2.4.0
soupsieve==2.4.1
SpeechRecognition==3.8.1
textract==1.6.5
tinycss2==1.2.1
tinytag==1.9.0
tk==0.1.0
tqdm==4.65.0
tzdata==2023.3
tzlocal==4.3
urllib3==1.26.15
virustotal-api==1.1.11
weasyprint==58.1
webencodings==0.5.1
xlrd==1.2.0
XlsxWriter==3.1.0
zopfli==0.2.2